This is what my sister and I did so far.
The completed project would be able to show stories from the Quraan, let people read the Quraan and let people listen to it on an mp3 version. It would also be able to recommend some chapters of the Quraan to read, based on what is trending. This was gonna be our AI-based function. We also designed a main page that is viewable on this link: https://www.w3schools.com/code/tryit.asp?filename=GI2C46AVO529